BREAKING: Complete client-server separation for OCR utilities
Resolves persistent Node.js module errors (dns, child_process, grpc) by:

1. MOVED client utilities to @repo/shared/utils/ocr-client.ts
   - ClientQuotaTracker, hasGoogleVisionApiKey, getOCRQuotaStatus
   - Zero server-side dependencies, pure browser-compatible code

2. UPDATED all imports to avoid @repo/ai/ocr entirely
   - ocr-settings.tsx: @repo/shared/utils/ocr-client
   - ocr-quota-display.tsx: @repo/shared/utils/ocr-client
   - Prevents loading google-vision.ts and gRPC dependencies

3. CLEANED exports and removed redundant files
   - Removed packages/ai/ocr/client.ts (consolidated)
   - Removed packages/shared/utils/ocr-client-utils.ts
   - Cleaned package.json exports in both packages

4. ARCHITECTURE: Complete separation achieved
   - Client: localStorage + API calls (no Node.js deps)
   - Server: OCR processing via /api/ocr route
   - Zero import overlap between client and server code

Fixes: dns, child_process, tls, net, http2 resolution errors
Tests: No compilation errors found in dry-run build
