+++ packages/ai/ocr/index.ts
@@ -3,9 +3,6 @@ export { TesseractService } from './tesseract';
 export { QuotaTracker } from './quota-tracker';
 export { OCRManager, getOCRManager } from './ocr-manager';
 
-// Client-safe exports
-export { ClientQuotaTracker, hasGoogleVisionApiKey } from './client';
-
 export type {
     OCRResult as GoogleVisionOCRResult,
     QuotaInfo,
