+++ packages/shared/utils/index.ts
@@ -1,7 +1,7 @@
 export * from './hotjar';
 export * from './messages';
 export * from './ocr-utils';
-export * from './ocr-client-utils';
+export * from './ocr-client';
 export * from './ocr-client-processing';
 export * from './plausible';
 export * from './tiptap-extensions';
